add templates folder and file structure
