+++ backend/templates/404.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="utf-8">
+    <title>Page Not Found :(</title>
+</head>
+
+<body>
+    <h1>Not found
+        <span>:(</span>
+    </h1>
+    <p>Sorry, but the page you were trying to view does not exist.</p>
+    <p>It looks like this was the result of either:</p>
+    <ul>
+        <li>a mistyped address</li>
+        <li>an out-of-date link</li>
+    </ul>
+</body>
+
+</html>
