+++ backend/templates/index.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+    <meta charset="utf-8">
+    <title>flask backend</title>
+    <meta name="description" content="Basic web app using flask + react">
+    <meta name="author" content="Riken Mehta">
+    <meta name="viewport" content="width=device-width , initial-scale=1, shrink-to-fit=no">
+    <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
+</head>
+
+<body>
+    <h1>This is the index.html</h1>
+</body>
+
+</html>
